requirements.txt
patool==1.12
XlsxWriter==3.0.3
